python/django-server/requirements.txt
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
boto3==1.4.4
botocore==1.5.39
decorator==4.0.11
Django==1.11
django-cors-middleware==1.3.1
djangorestframework==3.6.2
docutils==0.13.1
enum34==1.1.6
futures==3.1.1
ipython==5.3.0
ipython-genutils==0.2.0
jmespath==0.9.2
MySQL-python==1.2.5
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
s3transfer==0.1.10
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
utils==0.9.0
wcwidth==0.1.7
